+++ src/pages/Investment/InvestmentOverview.jsx
@@ -1,6 +1,6 @@
 import React, { useState, useMemo } from "react";
 import styles from "./InvestmentOverview.module.css";
-import CustomSelect from "./CustomSelect";
+import CustomSelect from "../../shared/component/CustomSelect";
 import InvestmentList from "./InvestmentList";
 import data from "./tempData";
 
@@ -10,6 +10,13 @@ function InvestmentOverview() {
   const handleSortOptionChange = (option) => {
     setSortOption(option);
   };
+
+  const sortOptions = [
+    { label: "View My Startup 투자 금액 높은순", value: "simInvest_desc" },
+    { label: "View My Startup 투자 금액 낮은순", value: "simInvest_asc" },
+    { label: "실제 누적 투자 금액 높은순", value: "actualInvest_desc" },
+    { label: "실제 누적 투자 금액 낮은순", value: "actualInvest_asc" },
+  ];
 
   const sortedData = useMemo(() => {
     return [...data].sort((a, b) => {
@@ -33,6 +40,7 @@ function InvestmentOverview() {
       <div className={styles.titleBar}>
         <div className={styles.title}>투자 현황</div>
         <CustomSelect
+          options={sortOptions} // options을 상위에서 전달
           onOptionChange={handleSortOptionChange}
           selectedOption={sortOption}
         />
